third_party/WebKit/LayoutTests/http/tests/security/vibration/vibrate-in-cross-origin-iframe-with-user-gesture-allowed-expected.txt

CONSOLE ERROR: line 19: Blocked call to navigator.vibrate inside a cross-origin iframe because the frame has never been activated by the user: https://www.chromestatus.com/feature/5682658461876224.
This tests that a cross-origin iframe with user gesture can vibrate.

127.0.0.1


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
This is a testharness.js-based test.
PASS Testing vibrate 3 times in an iframe
Harness: the test ran to completion.
